Template!
Incorporated the "Dashboard" template from the Bootstrap examples. I will work from this template and make adjustments as I go, putting in the different graphs and scripts.

I picked this template because, of the available ones, I believe it makes the most sense. It is simple to navigate between Sections I can rename, comes with a Search and Filter, and puts the graphs front and center.