Minor adjustments
A folder called "final_assets" was added to the app directory that contains the assets that will be used in the program. It includes images and already-processed CSV files. Finally, those images are being used as placeholders for the maps in the SSA category, so we can mark it as done!