Website successfully deployed!
With this update, the website is now working fully as intended in the Docs folder! It's been successfully deployed, and can be run using the index.html file. This change was made possible by switching to "HashRouter" in the React-Dom library, as opposed to the previous "BrowserRouter" that was being used, alongside some other changes to the "package.json" and Sidebar files. The next thing to do is polish the site further, mainly by using CSS.